ENH: Added copy of cns.channel to package
As this package is meant to eventually replace much of the core functionality
of the neurobehavior package that is generic enough to be used by other labs, I
am slowly copying relevant code over as I find a need for it.  I cleaned up the
file and applied PEP8 fixes before commiting, but did not change any
algorithms.