ENH: Updated expression evaluation system
All expressions must now be wrapped in an ExpressionNamespace object
which handles caching of values and getting the next value from
expressions that are generators.  As a consequence of this modification,
we can now indicate that an expression (defining a generator) returns a
new value *only* when another expression (also defining a generator)
raises a StopIteration.  As an example of what this allows us to do, we
can now loop through a sequence of levels.  Once that sequence is
exhaused, we shift to the next frequency and then loop through the
levels again.