ENH: Update abstract classes to use new exp eval
The abstract base classes (e.g. AbstractController) now use the new
expression evaluation system (i.e. ExpressionNamespace).  The eval
system has also been updated with new syntax to indicate when a
generator should be updated via an expression::

    u(<generator expression>, <dependency>)

This is sort of a hack, but should work.  Here, <generator expression>
defines the expression that returns a generator and <dependency>
indicates which expression to monitor for a StopIteration condition.
When the StopIteration condition is detected, the current generator's
next method is called.